+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGE LOG
 
+### v0.2.1
+
+- Feat: no need to manually call `init`. This function is removed.
+
 ### v0.2.0
 
 - Breaking: new method `init(incremental::Bool=true)` and `gen_scoped_stream_methods(incremental::Bool=true)`.